backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrs==25.4.0
autocommand==2.2.2
backports.tarfile==1.2.0
bce-python-sdk==0.9.59
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
cryptography==46.0.3
dashscope==1.25.7
duckduckgo-search
dataclasses-json==0.6.7
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
faiss-cpu==1.13.2
fastapi==0.110.0
filelock==3.20.2
FLAML==2.3.6
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
importlib-metadata==8.7.1
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.2
lxml==6.0.2
markupsafe==3.0.3
marshmallow==3.26.2
modelscope==1.33.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.15.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opt-einsum==3.3.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==12.1.0
pip==25.3
platformdirs==4.2.2
prettytable==3.17.0
primp==0.15.0
propcache==0.4.1
protobuf>=3.20.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyautogen==0.10.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pypdf==6.6.0
pypdfium2==5.3.0
pyreadline3==3.5.4
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.2
rapidocr-onnxruntime==1.2.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==3.3.0
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
starlette==0.36.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.0.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typeguard==4.3.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid-utils==0.13.0
uvicorn==0.27.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==16.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
